backend/Ecommerce_App/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0
idna==3.10
kombu==5.5.4
numpy==2.2.5
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
razorpay==1.4.2
redis==5.2.1
reportlab==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==80.3.1
six==1.17.0
social-auth-core==4.5.6
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
XlsxWriter==3.2.3
